requirements.txt
# =========================
# RAG + LLM
# =========================
llama-index==0.9.44
sentence-transformers==2.6.1
transformers==4.38.2
faiss-cpu==1.7.4
tiktoken==0.7.0

# =========================
# LOCAL LLM (Ollama)
# =========================
ollama==0.6.1

# =========================
# STREAMLIT UI
# =========================
streamlit==1.28.0

# =========================
# OCR + PDF + Images
# =========================
pytesseract==0.3.13
pdf2image==1.16.3
Pillow==10.0.1
pypdf>=6.1.3,<7
